requirements.txt
adal==1.2.6
applicationinsights==0.11.9
attrs==20.3.0
azure-common==1.1.26
azure-core==1.11.0
azure-graphrbac==0.61.1
azure-identity==1.4.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==2.2.0
azure-mgmt-resource==12.0.0
azure-mgmt-storage==11.2.0
azureml-core==1.23.0
azureml-dataprep==2.10.1
azureml-dataprep-native==30.0.0
azureml-dataprep-rslex==1.8.0
azureml-dataset-runtime==1.23.0
azureml-opendatasets==1.23.0
azureml-telemetry==1.23.0
backports.tempfile==1.0
backports.weakref==1.0.post1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
cloudpickle==1.6.0
contextlib2==0.6.0.post1
cryptography==3.2
distro==1.5.0
docker==4.4.3
dotnetcore2==2.1.20
Fiona==1.8.18
fusepy==3.0.1
geopandas==0.8.2
idna==2.10
isodate==0.6.0
jeepney==0.6.0
jmespath==0.10.0
jsonpickle==2.0.0
msal==1.9.0
msal-extensions==0.2.2
msrest==0.6.21
msrestazure==0.6.4
munch==2.5.0
ndg-httpsclient==0.5.1
numpy==1.19.0
oauthlib==3.1.0
pandas==1.0.0
pathspec==0.8.1
portalocker==1.7.1
py4j==0.10.9
pyarrow==1.0.1
pyasn1==0.4.8
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyproj==3.0.0.post1
pyspark==3.0.2
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
Rtree==0.9.7
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scipy==1.4.1
SecretStorage==3.3.1
Shapely==1.7.1
six==1.15.0
urllib3==1.26.3
websocket-client==0.57.0
